Update Firefly and the data importer to the latest stable versions
There are newer stable versions of both Firefly III and the data
importer. Also, he CSV importer of the current version of the data
importer (0.9.0) doesn't seem to work.

Update the version of Firefly III and the data importer in the
`composer` commands. We also need to update to PHP 8.2 in order to
update the data importer to the latest stable version (1.0.0).